Refactoring: #1 after mentor's review
