+++ src/Routes.js
@@ -3,8 +3,8 @@ import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
 import Login from './pages/Login/Login';
 import Main from './pages/Main/Main';
 import Products from './pages/Products/Products';
-import DetailMain from './pages/Products/DetailMain/DetailMain';
-import Category from './pages/Products/DetailMain/Category/Category';
+import Detail from './pages/Products/Detail';
+import Category from './pages/Products/Components/Category';
 
 class Routes extends Component {
   render() {
@@ -14,7 +14,7 @@ class Routes extends Component {
           <Route exact path="/" component={Main} />
           <Route exact path="/Login" component={Login} />
           <Route exact path="/Products" component={Products} />
-          <Route exact path="/Products/detail" component={DetailMain} />
+          <Route exact path="/Products/detail" component={Detail} />
           <Route exact path="/Products/category" component={Category} />
           {/* <Route exact path="/Products/:id" component={Detail} /> */}
         </Switch>
