+++ src/pages/Products/Components/Buttons.scss
@@ -1,12 +1,12 @@
-@import './../../../../styles/variables.scss';
+@import '../../../styles/variables.scss';
 
 .btns {
   @include flexBtw(space-between);
-  width: 580px;
 
   .btn {
     width: 285px;
     height: 50px;
+    margin: 5px;
     background-color: transparent;
     font-size: inherit;
     font-weight: bold;
