+++ src/pages/Products/Components/Card.scss
@@ -1,0 +1,59 @@
+@import '../../../styles/variables.scss';
+
+.card {
+  display: inline-block;
+  max-width: 285px;
+  margin: 10px;
+  text-align: center;
+
+  img {
+    width: 285px;
+    height: 285px;
+    object-fit: cover;
+  }
+
+  #tags {
+    display: inline-block;
+    width: 80px;
+    height: 23px;
+    padding: 3px;
+    margin: 5px 3px;
+    color: $color-white;
+    border-radius: 15px;
+    font-size: $fs-s;
+    font-weight: bold;
+    text-transform: uppercase;
+    line-height: 1.4em;
+
+    &.new {
+      background-color: #007929;
+    }
+
+    &.vegan {
+      background-color: #92ad00;
+    }
+
+    &.best {
+      background-color: $color-black;
+    }
+
+    &.soldOut {
+      background-color: #ff8a00;
+    }
+  }
+
+  .title {
+    margin-bottom: 5px;
+    font-weight: bold;
+  }
+
+  .hashtag {
+    margin-bottom: 20px;
+    color: grey;
+    font-size: 14px;
+  }
+
+  .price {
+    font-weight: bold;
+  }
+}
