+++ src/pages/Products/Components/Category.scss
@@ -1,4 +1,4 @@
-@import './../../../../styles/variables.scss';
+@import '../../../styles/variables.scss';
 
 .category {
   margin: 60px 0;
