+++ src/pages/Products/Components/DetailImages.scss
@@ -1,4 +1,4 @@
-@import './../../../../styles/variables.scss';
+@import '../../../styles/variables.scss';
 
 .detailImage {
   @include flexCenter(column);
