+++ src/pages/Products/Components/DetailInfo.scss
@@ -1,4 +1,4 @@
-@import './../../../../styles/variables.scss';
+@import '../../../styles/variables.scss';
 
 .detailInfo {
   position: relative;
