+++ src/pages/Products/Components/DetailNavi.js
@@ -1,0 +1,8 @@
+import React, { Component } from 'react';
+import './DetailNavi.scss';
+
+export default class DetailNavi extends Component {
+  render() {
+    return <div></div>;
+  }
+}
