+++ src/pages/Products/Components/DetailYoutube.js
@@ -1,0 +1,10 @@
+import React, { Component } from 'react';
+import './DetailYoutube.scss';
+
+class DetailYoutube extends Component {
+  render() {
+    return <div></div>;
+  }
+}
+
+export default DetailYoutube;
