+++ src/pages/Products/Components/Headers.scss
@@ -1,4 +1,4 @@
-@import './../../../styles/variables.scss';
+@import '../../../styles/variables.scss';
 
 header {
   position: relative;
