+++ src/pages/Products/Components/OrderInfo.scss
@@ -1,4 +1,4 @@
-@import './../../../../styles/variables.scss';
+@import '../../../styles/variables.scss';
 
 .orderInfo {
   @include flexBtw();
@@ -15,11 +15,15 @@
   }
 }
 
-.tag {
+.hashTags {
   position: relative;
   margin-bottom: 25px;
   color: #808080;
   font-weight: bold;
+
+  &.small {
+    font-size: $fs-m;
+  }
 }
 
 .infoWrap {
