+++ src/pages/Products/Detail.scss
@@ -1,4 +1,4 @@
-@import './../../../styles/variables.scss';
+@import './../../styles/variables.scss';
 
 main {
   @include flexCenter();
