+++ src/pages/Products/Products.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
-import Headers from './Headers/Headers';
-import Lists from './Lists/Lists';
+import Headers from './Components/Headers';
+import Lists from './Components/Lists';
 import './Products.scss';
 
 class Products extends Component {
