+++ src/pages/Products/Products.scss
@@ -1,19 +1,12 @@
 @import './../../styles/variables.scss';
 
-@mixin flexCenter($direction) {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-direction: $direction;
-}
-
 .products {
   width: 100vw;
-  height: 100vh;
   margin: 0 auto;
   overflow-x: hidden;
   -ms-overflow-style: none;
   scrollbar-width: none;
+  // cursor: url('/images/lush.png'), auto;
 
   .selectedOption {
     position: relative;
@@ -22,16 +15,16 @@
     margin: 0 auto;
     font-size: 28px;
     font-weight: bold;
-  }
 
-  .selectedOption::after {
-    content: '';
-    position: absolute;
-    bottom: -20px;
-    left: 0;
-    width: 100%;
-    height: 3px;
-    background-color: black;
+    &::after {
+      content: '';
+      position: absolute;
+      bottom: -20px;
+      left: 0;
+      width: 100%;
+      height: 3px;
+      background-color: black;
+    }
   }
 
   button {
